function_app/requirements.txt
azure-functions
azure-search-documents
openai
python-dotenv
